from float32 to float64

+++ api/term_analysis/similarity_analyzer.py
@@ -99,7 +99,7 @@ def _calculate_cosine_similarity(logger, workspace_bow):
     """
     # normalized and calculate cosine similarity
     logger.info('1')
-    workspace_bow = np.asarray(workspace_bow, np.float32)
+    workspace_bow = np.asarray(workspace_bow, np.float64)
     workspace_bow = workspace_bow / np.linalg.norm(workspace_bow, axis=1, keepdims=True)
     logger.info('2')
     cosine_similarity_matrix = workspace_bow.dot(np.transpose(workspace_bow))
